requirements.txt
opencv-python
numpy
PyYAML
scikit-image
matplotlib
requests